Adding GML reader and writer
Includes tests based on examples in GML documentation
and existing GraphML tests

It is not possible to round trip exactly from arbitary
GML files, properties about the graph and comments are lost.

Extracted LexicographicalComparator so writers can use the
same sorting.

As GML file ids must be integers but Graph id can be any object
we generate interger ids in the GMLWriter and store the blueprint
ids in a separate property - these can be used when loading from
a GML file by setting properties on the loader
(simmilat to GraphML reader)