Combined the two EventGraph implementations.
Unified the EventGraph implementations (EventGraph and EventTransactionalGraph) so that they both serve the same purpose.  When working in non-transactional model EventGraph will simply fire events in the buffer immediately following their inclusion.  If it is transactional then it will only fire the buffered events when there is a commit or shutdown.