Properly drop database between requests. #268
This is from Luca and allows the tests to run.  The tests still fail, because OrientDB doesn't support a "." in key names and GraphSail uses such a key.  If the key is changed then everything should work.  Awaiting an update from Josh on whether it is ok to do that or not.