Add missing graph.shutdown() in txn graph test
The method testVertexPropertiesOnPreTransactionCommit() in
TransactionalGraphTestSuite generated a test graph but did not call
shutdown() on it.  I think omitting shutdown() is an error, but I
could be mistaken.  Every other method in the file that generates a
graph appears to call shutdown().  If this method really isn't
supposed to call shutdown(), then a comment noting it might be helpful
to future readers.

I found this when trying to upgrade Titan from Tinkerpop Stack 2.3.0
to 2.4.0-SNAPSHOT.  Omitting shutdown() caused errors during test
cleanup using a BDB TransactionalGraph implementation in Titan.