Added a test that replicates the Rexster integration test environment.
The purpose of the test is to provide a way to sanity check blueprints operations in Rexster without having to execute Rexster integration tests themselves.  By no means is this a replacement for Rexster integration tests, but at least it helps debug problems around blueprints usage with thread pools.

Currently the test fails for orientdb and therefore the test is named as an "untest".