Update to Neo4j 2.2.1
Since the Neo4j API has changed substantially this
involved some significant modifications. The major one
is that `setCheckElementsInTransaction` is now a (deprecated)
no-op, since it no longer appears possible to determine in a
straightforward manner whether a node or relationship has been
deleted in a given transaction.

The good news is that the consistency between the graph and legacy
index data seems to have been improved, so the indices no longer seem
to return deleted nodes/rels.

**Note the following _breaking_ API change:**

The `nodeIsDeleted` and `relationshipIsDeleted` **public** methods on the
graph have been actually removed, since:

1. I can't figure out how to implement them
2. They were only used (internally) for checking if elements were in
   a transaction, which is no longer required
3. It's better that people fix code to not depend on these methods than
   for them to do something wrong/misleading/throw an exception

Additionally, fixed various test code.

A note on compatibility: I've tested this on a reasonably substantial
project and the only differences I noticed from 2.1.x involved order
of relationship iteration, which shouldn't be depended on anyway.